// frontend/src/api/generated/models/subscriptionResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Chain Risk Platform - BFF
 * BFF API for Chain Risk Platform
 * OpenAPI spec version: 1.0
 */
import type { SubscriptionResponseChannelType } from './subscriptionResponseChannelType';
import type { SubscriptionResponseChannelConfig } from './subscriptionResponseChannelConfig';

export interface SubscriptionResponse {
  id: number;
  userId: string;
  ruleId?: number;
  channelType: SubscriptionResponseChannelType;
  channelConfig: SubscriptionResponseChannelConfig;
  enabled: boolean;
  createdAt: string;
  updatedAt: string;
}