// frontend/src/api/generated/models/createSubscriptionDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Chain Risk Platform - BFF
 * BFF API for Chain Risk Platform
 * OpenAPI spec version: 1.0
 */
import type { CreateSubscriptionDtoChannelType } from './createSubscriptionDtoChannelType';
import type { CreateSubscriptionDtoChannelConfig } from './createSubscriptionDtoChannelConfig';

export interface CreateSubscriptionDto {
  /** User ID */
  userId: string;
  /** Rule ID (null for all rules) */
  ruleId?: number;
  /** Notification channel */
  channelType: CreateSubscriptionDtoChannelType;
  /** Channel config (email, webhook url, etc) */
  channelConfig: CreateSubscriptionDtoChannelConfig;
  /** Enable subscription */
  enabled?: boolean;
}