// frontend/src/api/generated/models/createSubscriptionDtoChannelType.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Chain Risk Platform - BFF
 * BFF API for Chain Risk Platform
 * OpenAPI spec version: 1.0
 */

/**
 * Notification channel
 */
export type CreateSubscriptionDtoChannelType = typeof CreateSubscriptionDtoChannelType[keyof typeof CreateSubscriptionDtoChannelType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateSubscriptionDtoChannelType = {
  email: 'email',
  webhook: 'webhook',
  slack: 'slack',
  telegram: 'telegram',
} as const;